Pedro's Sketches
My version of the sketches with a slight change in the project, instead of a coffee maker presented, I have created a traffic light controller. The program works exactly like the main one, clicking the buttons will turn lights on and off on the board based on the message received by the server through the websocket.